+++ pyxelrest/fast_deserializer.py
@@ -26,6 +26,9 @@ def to_date_time(value):
         # Conversion cannot be performed for dates after year 3000, best effort and return in provided timezone
         if datetime_with_service_timezone.year > 3000:
             return datetime_with_service_timezone
+        # Conversion cannot be performed for dates <= to 1970-01-01 best effort and return in provided timezone
+        if datetime_with_service_timezone.year < 1971:
+            return datetime_with_service_timezone
         return datetime_with_service_timezone.astimezone(tz=dateutil.tz.tzlocal())
     return value