+++ pyxelrest_update_services_config.py
@@ -95,8 +95,6 @@ class ServicesConfigUpdater:
         if LIST_SECTIONS != self._action:
             logger.info('Updating services configuration...')
 
-        # TODO Remove once every client will have this version in conf
-        file_or_directory = file_or_directory.replace('all_services.ini', 'services.yml')
         updated_config = open_config(file_or_directory)
         if not updated_config:
             logger.error('Services configuration cannot be updated.')