+++ testsutils/non_authenticated_service.py
@@ -22,11 +22,12 @@ def open_api_definition():
 
 @app.route('/without_auth', methods=['GET'])
 def get_without_auth():
-    return jsonify([{'received token': request.headers.get('Bearer') is not None}])
+    return jsonify([{'received token': request.headers.get('Authorization') is not None}])
 
 
 def start_server(port):
     app.run(port=port)
+
 
 if __name__ == '__main__':
     start_server(8948)