+++ CHANGELOG.md
@@ -15,6 +15,7 @@ List all changes in various categories:
 - Allow to specify header name and header value for the non-standard OAuth2 services.
 - Logs are now identifying the Book, Sheet and Cell performing the call.
 - Document value 303 in wait_for_status pyxelrest_get_url parameter.
+- Rely on latest official xlwings release (0.13.0).
 
 ### Bug fixes ###