+++ CHANGELOG.md
@@ -28,7 +28,8 @@ List all changes in various categories:
 - Previous results are now cleared when using asynchronous UDFs with shifted results.
 - Make sure null values are sent in array only if swagger explicitly state that null values are allowed.
 - Handle date and date-time prior or equal to 1970-01-01T01:00:00+00:00
-- Support for pip > 9 (shipped with recent Python versions)
+- Support for pip > 9 (shipped with recent Python versions).
+- Handle more than one leading underscore in field name.
 
 ## 0.68.4 (2018-08-07) ##